requirements.txt
streamlit
pandas
groq
langchain
langchain-groq
langchain-community
langchain-huggingface
sentence-transformers
chromadb
python-dotenv